refactor(dashboard): extract mergeRouteRecord + hasRouteContext to normalize.ts
logs.ts and keeper-tool-call-inspector.ts had byte-identical
mergeRouteRecord (idString-based field extraction with overwrite guard)
and hasRouteContext (11-field undefined check). Extracted both to
normalize.ts with MutableRouteContext interface. Removed 2 local type
hierarchies (ToolRouteContextFields/MutableToolRouteContext,
LogRouteContextFields/MutableLogRouteContext) and 4 local functions.

session-trace-entry.ts keeps its own mergeTraceRouteRecord because it
uses firstStringish (multi-key, camelCase, array fallback) instead of
idString (single-key).
