+++ dashboard/src/components/keeper-tool-call-inspector.ts
@@ -8,7 +8,7 @@ import { fetchKeeperToolCalls } from '../api/dashboard'
 import type { ToolCallEntry, ToolCallsResponse, TelemetryFreshnessMetadata } from '../api/dashboard'
 import { formatTimeHms } from '../lib/format-time'
 import { LoadingState } from './common/feedback-state'
-import { asRecord, idString, extractCodeLocation } from './common/normalize'
+import { asRecord, mergeRouteRecord, hasRouteContext, type MutableRouteContext } from './common/normalize'
 import { SectionCap } from './common/section-cap'
 import { toolCategory, formatDuration, durationColor } from './tool-call-shared'
 import { useManagedAsyncResource } from '../lib/use-managed-async-resource'
@@ -20,7 +20,6 @@ import { coverageGapDisplay, sourceHealthClass, freshnessText } from './common/s
 import {
   openIdeContextRouteLink,
   routeLinksForContext,
-  type IdeContextRouteContext,
   type IdeContextRouteLink,
 } from './ide/ide-context-lens'
 
@@ -68,25 +67,6 @@ function tryPrettyJson(s: string): string | null {
   }
 }
 
-type ToolRouteContextFields = Pick<
-  IdeContextRouteContext,
-  | 'filePath'
-  | 'line'
-  | 'goalId'
-  | 'taskId'
-  | 'boardPostId'
-  | 'commentId'
-  | 'prId'
-  | 'gitRef'
-  | 'logId'
-  | 'sessionId'
-  | 'operationId'
-  | 'workerRunId'
->
-type MutableToolRouteContext = {
-  -readonly [K in keyof ToolRouteContextFields]?: ToolRouteContextFields[K]
-}
-
 function parseInputRecord(input: string): Record<string, unknown> | null {
   try {
     return asRecord(JSON.parse(input))
@@ -95,40 +75,8 @@ function parseInputRecord(input: string): Record<string, unknown> | null {
   }
 }
 
-function mergeToolRouteRecord(
-  context: MutableToolRouteContext,
-  record: Record<string, unknown> | null,
-  overwrite = false,
-): void {
-  if (!record) return
-  const location = extractCodeLocation(record)
-  if (location?.filePath && (overwrite || context.filePath === undefined)) context.filePath = location.filePath
-  if (location?.line !== undefined && (overwrite || context.line === undefined)) context.line = location.line
-
-  const goalId = idString(record.goal_id)
-  if (goalId && (overwrite || context.goalId === undefined)) context.goalId = goalId
-  const taskId = idString(record.task_id)
-  if (taskId && (overwrite || context.taskId === undefined)) context.taskId = taskId
-  const boardPostId = idString(record.board_post_id) ?? idString(record.post_id)
-  if (boardPostId && (overwrite || context.boardPostId === undefined)) context.boardPostId = boardPostId
-  const commentId = idString(record.comment_id) ?? idString(record.reply_id) ?? idString(record.comment_number)
-  if (commentId && (overwrite || context.commentId === undefined)) context.commentId = commentId
-  const prId = idString(record.pr_id) ?? idString(record.pull_request) ?? idString(record.pr_number)
-  if (prId && (overwrite || context.prId === undefined)) context.prId = prId
-  const gitRef = idString(record.git_ref) ?? idString(record.commit) ?? idString(record.branch)
-  if (gitRef && (overwrite || context.gitRef === undefined)) context.gitRef = gitRef
-  const logId = idString(record.log_id)
-  if (logId && (overwrite || context.logId === undefined)) context.logId = logId
-  const sessionId = idString(record.session_id)
-  if (sessionId && (overwrite || context.sessionId === undefined)) context.sessionId = sessionId
-  const operationId = idString(record.operation_id)
-  if (operationId && (overwrite || context.operationId === undefined)) context.operationId = operationId
-  const workerRunId = idString(record.worker_run_id)
-  if (workerRunId && (overwrite || context.workerRunId === undefined)) context.workerRunId = workerRunId
-}
-
 function mergeToolInputContext(
-  context: MutableToolRouteContext,
+  context: MutableRouteContext,
   input: unknown,
   depth = 0,
 ): void {
@@ -140,32 +88,18 @@ function mergeToolInputContext(
   const record = asRecord(input)
   if (!record) return
   const failureEnvelope = asRecord(record.failure_envelope)
-  mergeToolRouteRecord(context, asRecord(record.context))
-  mergeToolRouteRecord(context, asRecord(record.evidence_ref))
-  mergeToolRouteRecord(context, asRecord(failureEnvelope?.evidence_ref))
-  mergeToolRouteRecord(context, asRecord(record.tool_args))
+  mergeRouteRecord(context, asRecord(record.context))
+  mergeRouteRecord(context, asRecord(record.evidence_ref))
+  mergeRouteRecord(context, asRecord(failureEnvelope?.evidence_ref))
+  mergeRouteRecord(context, asRecord(record.tool_args))
   mergeToolInputContext(context, record.input, depth + 1)
-  mergeToolRouteRecord(context, record, true)
-}
-
-function hasToolRouteContext(context: MutableToolRouteContext): boolean {
-  return context.filePath !== undefined
-    || context.goalId !== undefined
-    || context.taskId !== undefined
-    || context.boardPostId !== undefined
-    || context.commentId !== undefined
-    || context.prId !== undefined
-    || context.gitRef !== undefined
-    || context.logId !== undefined
-    || context.sessionId !== undefined
-    || context.operationId !== undefined
-    || context.workerRunId !== undefined
+  mergeRouteRecord(context, record, true)
 }
 
 function toolCallRouteLinks(entry: ToolCallEntry): ReadonlyArray<IdeContextRouteLink> {
-  const context: MutableToolRouteContext = {}
+  const context: MutableRouteContext = {}
   mergeToolInputContext(context, entry.input)
-  if (!hasToolRouteContext(context)) return []
+  if (!hasRouteContext(context)) return []
   const links = routeLinksForContext({
     ...context,
     surface: 'Tool',
